tpm2_create: refactor object and session interface
The Eys_Create and Esys_CreateLoaded both take:
1. A handle to the parent object
2. 3 session handles, with the first session handle relating to
   the auth for the parent object.

Rather than have the tool:
1. call Esys_TR_SetAuth() and populate shandle1
2. manage the parent and it's auth seperately

Just refactor the tpm2.c interface to take a tpm2_loaded_object *
for the parent and move a tpm2_session * into the tpm2_loaded_object.
Now Auth and Object live together, thus correcting issue 2. Inside
of tpm2.c Esys_ creation routines, just get the shandle1 and the
esys tr handle from the loaded object.

This makes the command interface for the tpm from tools much simpler,
and removes complexity from the tools. This paradigm can be applied
to all the tools, and eventually tpm2_util_object_load() can probably
go away.

Progresses: #1193
Progresses: #1539
