+++ tools/misc/tpm2_checkquote.c
@@ -11,6 +11,7 @@
 
 #include "files.h"
 #include "log.h"
+#include "object.h"
 #include "pcr.h"
 #include "tpm2_alg_util.h"
 #include "tpm2_convert.h"
