+++ tools/tpm2_activatecredential.c
@@ -14,6 +14,7 @@
 
 #include "files.h"
 #include "log.h"
+#include "object.h"
 #include "tpm2_auth_util.h"
 #include "tpm2_error.h"
 #include "tpm2_options.h"
