+++ tools/tpm2_certify.c
@@ -10,6 +10,7 @@
 
 #include "files.h"
 #include "log.h"
+#include "object.h"
 #include "tpm2.h"
 #include "tpm2_alg_util.h"
 #include "tpm2_auth_util.h"
