+++ tools/tpm2_createak.c
@@ -14,6 +14,7 @@
 #include "tpm2_auth_util.h"
 #include "files.h"
 #include "log.h"
+#include "object.h"
 #include "tpm2_util.h"
 #include "tpm2_session.h"
 #include "tpm2_alg_util.h"
