+++ tools/tpm2_duplicate.c
@@ -7,6 +7,7 @@
 #include <tss2/tss2_esys.h>
 
 #include "log.h"
+#include "object.h"
 #include "files.h"
 #include "tpm2_alg_util.h"
 #include "tpm2_auth_util.h"
