+++ tools/tpm2_encryptdecrypt.c
@@ -12,6 +12,7 @@
 
 #include "files.h"
 #include "log.h"
+#include "object.h"
 #include "tpm2.h"
 #include "tpm2_alg_util.h"
 #include "tpm2_auth_util.h"
