+++ tools/tpm2_policysecret.c
@@ -6,6 +6,7 @@
 
 #include "files.h"
 #include "log.h"
+#include "object.h"
 #include "tpm2_auth_util.h"
 #include "tpm2_hierarchy.h"
 #include "tpm2_options.h"
