+++ tools/tpm2_readpublic.c
@@ -9,6 +9,7 @@
 #include "tpm2_convert.h"
 #include "files.h"
 #include "log.h"
+#include "object.h"
 #include "tpm2.h"
 #include "tpm2_alg_util.h"
 #include "tpm2_attr_util.h"
