+++ tools/tpm2_startauthsession.c
@@ -13,6 +13,7 @@
 
 #include "files.h"
 #include "log.h"
+#include "object.h"
 #include "tpm2_alg_util.h"
 #include "tpm2_options.h"
 #include "tpm2_session.h"
