+++ tools/tpm2_unseal.c
@@ -9,6 +9,7 @@
 
 #include "files.h"
 #include "log.h"
+#include "object.h"
 #include "pcr.h"
 #include "tpm2_auth_util.h"
 #include "tpm2_hash.h"
