+++ tools/tpm2_verifysignature.c
@@ -8,6 +8,7 @@
 
 #include "files.h"
 #include "log.h"
+#include "object.h"
 #include "tpm2_alg_util.h"
 #include "tpm2_convert.h"
 #include "tpm2_hash.h"
